feat! AIM cli
* creates student_workers subcommand with expire_passwords action
* creates thor cli
