+++ Dockerfile
@@ -1,5 +1,4 @@
-ARG RUBY_VERSION=3.2
-FROM ruby:${RUBY_VERSION}
+FROM ruby:3.2 AS development
 
 ARG UNAME=app
 ARG UID=1000
@@ -16,9 +15,17 @@ RUN groupadd -g ${GID} -o ${UNAME}
 RUN useradd -m -d /app -u ${UID} -g ${GID} -o -s /bin/bash ${UNAME}
 RUN mkdir -p /gems && chown ${UID}:${GID} /gems
 
-
+ENV PATH="$PATH:/app/exe:/app/bin"
 USER $UNAME
 
 ENV BUNDLE_PATH /gems
 
 WORKDIR /app
+
+FROM development AS production
+
+COPY --chown=${UID}:${GID} . /app
+
+RUN --mount=type=secret,id=gh_package_read_token,uid=1000 \
+  read_token="$(cat /run/secrets/gh_package_read_token)" \
+  && BUNDLE_RUBYGEMS__PKG__GITHUB__COM=${read_token} bundle install
