--- Dockerfile.prod
@@ -1,31 +1,0 @@
-ARG RUBY_VERSION=3.2
-FROM ruby:${RUBY_VERSION}
-
-ARG UNAME=app
-ARG UID=1000
-ARG GID=1000
-
-LABEL maintainer="mrio@umich.edu"
-
-RUN apt-get update -yqq && apt-get install -yqq --no-install-recommends \
-  vim-tiny
-
-RUN gem install bundler
-
-RUN groupadd -g ${GID} -o ${UNAME}
-RUN useradd -m -d /app -u ${UID} -g ${GID} -o -s /bin/bash ${UNAME}
-RUN mkdir -p /gems && chown ${UID}:${GID} /gems
-
-ENV BUNDLE_PATH /gems
-
-COPY --chown=${UID}:${GID} . /app
-
-WORKDIR /app
-
-RUN --mount=type=secret,id=gh_package_read_token \
-  read_token="$(cat /run/secrets/gh_package_read_token)" \
-  && BUNDLE_RUBYGEMS__PKG__GITHUB__COM=${read_token} bundle install
-
-USER $UNAME
-
-CMD ["bundle", "exec", "ruby", "./expire_passwords.rb"]
