+++ docker-compose.yml
@@ -1,8 +1,10 @@
-version: '3'
+version: '3.8'
 
 services:
-  web:
-    build: .
+  app:
+    build:
+      context: .
+      target: development
     volumes:
       - .:/app
       - gem_cache:/gems
