+++ exe/aim
@@ -1,0 +1,5 @@
+#!/usr/bin/env ruby
+#require "aim"
+require "/app/lib/aim"
+
+AIM::CLI.start
