+++ lib/aim.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require "thor"
+require_relative "aim/student_workers"
+require_relative "aim/cli"
+
+module AIM
+  class Error < StandardError; end
+  # Your code goes here...
+end
