+++ lib/aim/student_workers.rb
@@ -1,0 +1,9 @@
+require "alma_rest_client"
+require "logger"
+require "byebug"
+require_relative "student_workers/password_expirer"
+
+module AIM
+  module StudentWorkers
+  end
+end
