+++ spec/fixtures/student_workers/set_members.json
